requirements.txt
Flask==2.3.2
numpy==1.23.5
tensorflow-cpu==2.12.0
scikit-learn==1.2.2
pandas==1.5.3
Pillow==9.5.0
imbalanced-learn==0.10.1
opencv-python==4.7.0.72
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4